\name{add_zone}
\alias{add_zone}

\title{
  Adds a strike zone layer to current graph
}
\description{
  Adds a strike zone layer to current graph
}
\usage{
  add_zone(Color = "red")
}
\arguments{
  \item{Color}{
    color of zone edge
  }
}
\value{
  ggplot2 layer to add a zone to current graph
}
\author{
  Jim Albert
}


