\name{called_strike_plot}
\alias{called_strike_plot}

\title{
  Called Strike Plot
}
\description{
  Constructs called strike plot for a specific player
}
\usage{
  called_strike_plot(data, title)
}
\arguments{
  \item{data}{
    data frame containing Statcast data
  }
  \item{title}{
    title for the graph
  }
}
\value{
  Constructs a tile plot of the called strike probability
}
\author{
  Jim Albert
}


