\name{collect_4_years}
\alias{collect_4_years}

\title{
  Collects Statcast data for four years
}
\description{
  Collects Statcast data for years 2015 through 2018 for a particular player
}
\usage{
  collect_4_years(player)
}
\arguments{
  \item{player}{
    name of player
  }
}
\value{
  list of Statcast datasets for player for four seasons
}
\author{
  Jim Albert
}


