\name{collect_many_years}
\alias{collect_many_years}

\title{
  Collects Statcast data for one player for many seasons
}
\description{
  Collects Statcast data for one batter for many seasons
}
\usage{
  collect_many_years(player, years)
}
\arguments{
  \item{player}{
    name of player
  }
  \item{years}{
   vector of seasons of interest
  }
}
\value{
  Statcast dataset for player for many seasons
}
\author{
  Jim Albert
}


