\name{compute_pitch_values}
\alias{compute_pitch_values}

\title{
  Pitch Value Computation
}
\description{
  Computes pitch values for all pitches in a Statcast dataset
}
\usage{
  compute_pitch_values(sc)
}
\arguments{
  \item{sc}{
    Statcast data frame
  }
}
\value{
  Returns Statcast data frame with new variables Count, Runs.x, new_Count, Runs.y and Pitch_Value
}
\author{
  Jim Albert
}


