\name{contact_swing_contour}
\alias{contact_swing_contour}

\title{
  Contact on Swings Contour Plot
}
\description{
  Constructs contact on swung balls contour graph
}
\usage{
  contact_swing_contour(data, L, title)
}
\arguments{
  \item{data}{
    data frame containing Statcast data
  }
  \item{L}{
    values of contour values
  }
  \item{title}{
    title of the graph
  }
}
\value{
  Constructs a contour plot of the contact on swung balls probability
}
\author{
  Jim Albert
}


