\name{contact_swing_plot}
\alias{contact_swing_plot}

\title{
  Contact on Swings Plot
}
\description{
  Constructs contact on swung balls graph
}
\usage{
  contact_swing_plot(data, title)
}
\arguments{
  \item{data}{
    data frame containing Statcast data
  }
  \item{title}{
    title for the graph
  }
}
\value{
  Constructs a tile plot of the contact on swung balls probability
}
\author{
  Jim Albert
}


