\name{count_values}
\alias{count_values}

\title{
  Count run values from 2019 data
}
\description{
  Run values of all counts and plays from 2019 data
}
\usage{
  count_values
}
\format{
   A data frame with 45 observations on the following 2 variables.
  \describe{
  \item{Count}{count or end of PA event}
  \item{Runs}{average runs value}
  }
}
\source{
Retrosheet and Baseball Savant
}


