\name{cplot2}
\alias{cplot2}

\title{
  Constructs heat map plot from gam fit
}
\description{
  Constructs heat map plot from output from an generalized additive fit
}
\usage{
  cplot2(fit)
}
\arguments{
  \item{fit}{
    saved gam fit with binary response
  }
}
\value{
  ggplot2 heat map graph with strike zone added
}
\author{
  Jim Albert
}


