\name{ewoba_plot}
\alias{ewoba_plot}

\title{
  Expected wOBA Plot
}
\description{
  Constructs expected wOBA plot for a specific player
}
\usage{
  ewoba_plot(df, title = "Expected wOBA", NCOL = 2)
}
\arguments{
  \item{df}{
    data frame or list containing Statcast data
  }
  \item{title}{
    title for the graph
  }
  \item{NCOL}{
    number of columns in multipanel display
  }
}
\value{
  Constructs a tile plot of the expected wOBA
}
\author{
  Jim Albert
}


