\name{grid_predict}
\alias{grid_predict}

\title{
 Creates zone grid and predicted values
}
\description{
  Creates a data frame of zone and predicted values.
}
\usage{
  grid_predict(fit)
}
\arguments{
  \item{fit}{
    result of a GAM fit
  }
}
\value{
  data frame of the grid and predicted values
}
\author{
  Jim Albert
}


