\name{home_run_contour}
\alias{home_run_contour}

\title{
  Home Run Contour Plot
}
\description{
  Constructs home run contour plot for a specific player
}
\usage{
  home_run_contour(data, L, title)
}
\arguments{
  \item{data}{
    data frame containing Statcast data
  }
  \item{L}{
    values for the contour lines
  }
  \item{title}{
    title of the graph
  }
}
\value{
  Constructs a contour plot of the home run probability
}
\author{
  Jim Albert
}


