\name{inplay_swing_plot}
\alias{inplay_swing_plot}

\title{
  In-Play on Swings Plot
}
\description{
  Constructs in-play on swings graph
}
\usage{
  inplay_swing_plot(df, title = "Probability of In-Play", NCOL = 2)
}
\arguments{
  \item{df}{
    data frame or list containing Statcast data
  }
  \item{title}{
    title for the graph
  }
  \item{NCOL}{
    number of columns in multipanel display
  }
}
\value{
  Constructs a tile plot of the in-play on swung balls probability
}
\author{
  Jim Albert
}


