\name{location_compare}
\alias{location_compare}

\title{
  Pitch Location Graph
}
\description{
  Constructs pitch location graph
}
\usage{
  location_compare(df, title = "", NCOL = 2)
}
\arguments{
  \item{df}{
    data frame or list containing Statcast data
  }
  \item{title}{
    title for the graph
  }
  \item{NCOL}{
    number of columns in multipanel display
  }
}
\value{
  Constructs a density estimate of the pitch locations
}
\author{
  Jim Albert
}


