\name{ls_contour}
\alias{ls_contour}

\title{
  Launch Speed Contour Plot
}
\description{
  Constructs launch speed contour graph
}
\usage{
  ls_contour(data, L, title)
}
\arguments{
  \item{data}{
    data frame containing Statcast data
  }
  \item{L}{
    values for contour lines
  }
  \item{title}{
    title of the graph
  }
}
\value{
  Constructs a contour plot of the launch speeds
}
\author{
  Jim Albert
}


