\name{ls_plot}
\alias{ls_plot}

\title{
  Launch Speed Plot
}
\description{
  Constructs launch speed graph
}
\usage{
  ls_plot(df, title = "Launch Velocity", NCOL = 2)
}
\arguments{
  \item{df}{
    data frame or list containing Statcast data
  }
  \item{title}{
    title for the graph
  }
  \item{NCOL}{
    number of columns in multipanel display
  }
}
\value{
  Constructs a tile plot of the launch speeds
}
\author{
  Jim Albert
}


