\name{miss_swing_plot}
\alias{miss_swing_plot}

\title{
  Miss on Swings Plot
}
\description{
  Constructs miss on swung balls graph
}
\usage{
  miss_swing_plot(data, title)
}
\arguments{
  \item{data}{
    data frame containing Statcast data
  }
  \item{title}{
    title for the graph
  }
}
\value{
  Constructs a tile plot of the miss on swung balls probability
}
\author{
  Jim Albert
}


