\name{pitchers_2019}
\alias{pitchers_2019}

\title{
  Statcast Ids for 20 Pitchers
}
\description{
  Statcast ids for 20 pitchers
}
\usage{
  sc_pitchers
}
\format{
   A data frame with 20 observations on the following 2 variables.
  \describe{
  \item{Name}{name of the pitcher}
  \item{SC_id}{pitcher's Statcast id}
  }
}
\source{
Baseball Savant
}


