\name{sa_plot}
\alias{sa_plot}

\title{
  Spray Angle Plot
}
\description{
  Constructs spray angle graph
}
\usage{
  sa_plot(data, title)
}
\arguments{
  \item{data}{
    data frame containing Statcast data
  }
  \item{title}{
    title for the graph
  }
}
\value{
  Constructs a tile plot of the spray angles
}
\author{
  Jim Albert
}


