\name{sc2019_ip}
\alias{sc2019_ip}

\title{
  2019 Statcast inplay data
}
\description{
  Statcast inplay data for the 2019 season
}
\usage{
  sc2019_ip
}
\format{
   A data frame with 123825 observations on the following 4 variables.
  \describe{
  \item{player_name}{name of hitter}
  \item{plate_x}{horizontal position of ball when it crosses plate from catcher's perspective}
  \item{plate_z}{vertical position of ball when it crosses plate from catcher's perspective}
  \item{H}{logical variable if hit was recorded}
  }
}
\source{
Baseball Savant
}


