\name{select_sc_var}
\alias{select_sc_var}

\title{
  Selects subset of variables from a Statcast data frame
}
\description{
  Selects subset of variables from a Statcast data frame that are relevant for this package
}
\usage{
  select_sc_var(sc)
}
\arguments{
  \item{sc}{
    Statcast data fram
  }
}
\value{
  subset of Statcast data containing relevant variables
}
\author{
  Jim Albert
}


