\name{setup_all}
\alias{setup_all}

\title{
  Sets up Statcast data to define new variables
}
\description{
  Sets up Statcast data to define Swing and Count
}
\usage{
  setup_all(sc)
}
\arguments{
  \item{sc}{
    Statcast data frame
  }
}
\value{
  Statcast data frame for all pitches with new variables Swing and Count
}
\author{
  Jim Albert
}


