\name{setup_called}
\alias{setup_called}

\title{
  Sets up Statcast data to obtain called pitches
}
\description{
  Sets up Statcast data to obtain called pitches
}
\usage{
  setup_called(d)
}
\arguments{
  \item{d}{
    Statcast data frame
  }
}
\value{
  Statcast data frame for only the called pitches with new variables Strike and Count
}
\author{
  Jim Albert
}


