\name{setup_inplay}
\alias{setup_inplay}

\title{
  Sets up Statcast data to obtain inplay pitches
}
\description{
  Sets up Statcast data to obtain inplay pitches
}
\usage{
  setup_inplay(d)
}
\arguments{
  \item{d}{
    Statcast data frame
  }
}
\value{
  Statcast data frame for only the inplay pitches with new variables H and HR
}
\author{
  Jim Albert
}


