\name{setup_LR}
\alias{setup_LR}

\title{
  Partitions Statcast data frame by pitcher side
}
\description{
  Partitions Statcast data frame by pitcher side
}
\usage{
  setup_LR(sc)
}
\arguments{
  \item{sc}{
    Statcast data frame
  }
}
\value{
  List of two elements dividing data frame by pitcher side
}
\author{
  Jim Albert
}


