\name{setup_LR_pitchtype}
\alias{setup_LR_pitchtype}

\title{
  Partitions Statcast data frame by pitcher side and pitch type
}
\description{
  Partitions Statcast data frame by pitcher side and pitch type
}
\usage{
  setup_LR_pitchtype(sc)
}
\arguments{
  \item{sc}{
    Statcast data frame
  }
}
\value{
  List of four elements dividing data frame by pitcher side and pitch type
}
\author{
  Jim Albert
}


