\name{setup_pitchtype}
\alias{setup_pitchtype}

\title{
  Partitions Statcast data frame by and pitch type
}
\description{
  Partitions Statcast data frame by pitch type
}
\usage{
  setup_pitchtype(sc)
}
\arguments{
  \item{sc}{
    Statcast data frame
  }
}
\value{
  List of two elements dividing data frame by pitch type
}
\author{
  Jim Albert
}


