\name{setup_swing}
\alias{setup_swing}

\title{
  Sets up Statcast data to obtain swing pitches
}
\description{
  Sets up Statcast data to obtain swing pitches
}
\usage{
  setup_swing(d)
}
\arguments{
  \item{d}{
    Statcast data frame
  }
}
\value{
  Statcast data frame for only the swing pitches with new variables Miss and Count
}
\author{
  Jim Albert
}


