\name{split_count}
\alias{split_count}

\title{
  Partitions Statcast data frame by count
}
\description{
  Partitions Statcast data frame by count type
}
\usage{
  split_count(sc)
}
\arguments{
  \item{sc}{
    Statcast data frame
  }
}
\value{
  List of three elements dividing data frame by count (Ahead, Neutral and Behind)
}
\author{
  Jim Albert
}


