\name{split_LR}
\alias{split_LR}

\title{
  Partitions Statcast data frame by pitcher side
}
\description{
  Partitions Statcast data frame by pitcher side
}
\usage{
  split_LR(sc)
}
\arguments{
  \item{sc}{
    Statcast data frame
  }
}
\value{
  List of two elements dividing data frame by pitcher side
}
\author{
  Jim Albert
}


