\name{tile_plot_m}
\alias{tile_plot_m}

\title{
  Constructs heat map plot from gam fit
}
\description{
  Constructs heat map plot from output from an generalized additive fit
}
\usage{
  tile_plot_m(fit)
}
\arguments{
  \item{fit}{
    saved gam fit with continuous response
  }
}
\value{
  ggplot2 heat map graph with strike zone added
}
\author{
  Jim Albert
}


