\name{collect_player}
\alias{collect_player}

\title{
  Collects Statcast data for one player
}
\description{
  Collects Statcast data for one player for one season
}
\usage{
  collect_player(player, Season = 2018)
}
\arguments{
  \item{player}{
    name of player
  }
  \item{Season}{
   season of interest
  }
}
\value{
  Statcast dataset for player for one seasons
}
\author{
  Jim Albert
}


