\name{contour_graph}
\alias{contour_graph}

\title{
  Contour Graph
}
\description{
  Constructs filled contour graph
}
\usage{
  contour_graph(fit, L, title)
}
\arguments{
  \item{fit}{
    object from a GAM fit
  }
  \item{L}{
    vector of values for contour lines
  }
  \item{title}{
    title for the graph
  }
}
\value{
  Constructs a filled contour graph
}
\author{
  Jim Albert
}


