\name{ehit_plot}
\alias{ehit_plot}

\title{
  Expected Hit Plot
}
\description{
  Constructs expected hit plot for a specific player
}
\usage{
  ehit_plot(df, title = "Expected Batting Average")
}
\arguments{
  \item{df}{
    data frame or list containing Statcast data
  }
  \item{title}{
    title for the graph
  }
}
\value{
  Constructs a tile plot of the expected batting average
}
\author{
  Jim Albert
}


