\name{ewoba_plot}
\alias{ewoba_plot}

\title{
  Expected wOBA Plot
}
\description{
  Constructs expected wOBA plot for a specific player
}
\usage{
  ewoba_plot(df, title = "Expected wOBA")
}
\arguments{
  \item{df}{
    data frame or list containing Statcast data
  }
  \item{title}{
    title for the graph
  }
}
\value{
  Constructs a tile plot of the expected wOBA
}
\author{
  Jim Albert
}


