\name{hit_contour}
\alias{hit_contour}

\title{
  Hit Contour Plot
}
\description{
  Constructs hit contour plot for a specific player
}
\usage{
  hit_contour(data, L)
}
\arguments{
  \item{data}{
    data frame containing Statcast data
  }
  \item{L}{
    values for contour lines
  }
}
\value{
  Constructs a contour plot of the hit probability
}
\author{
  Jim Albert
}


