\name{hit_plot}
\alias{hit_plot}

\title{
  Hit Plot
}
\description{
  Constructs hit plot for a specific player
}
\usage{
  hit_plot(df, title = "Probability of a Hit")
}
\arguments{
  \item{df}{
    data frame or list containing Statcast data
  }
  \item{title}{
    title for the graph
  }
}
\value{
  Constructs a tile plot of the hit probability
}
\author{
  Jim Albert
}


