\name{home_run_plot}
\alias{home_run_plot}

\title{
  Home Run Plot
}
\description{
  Constructs home run plot for a specific player
}
\usage{
  home_run_plot(data, title)
}
\arguments{
  \item{data}{
    data frame containing Statcast data
  }
  \item{title}{
    title for the graph
  }
}
\value{
  Constructs a tile plot of the home run probability
}
\author{
  Jim Albert
}


