\name{la_contour}
\alias{la_contour}

\title{
  Launch Angle Contour Plot
}
\description{
  Constructs launch angle contour graph
}
\usage{
  la_contour(data, L)
}
\arguments{
  \item{data}{
    data frame containing Statcast data
  }
  \item{L}{
    values for contour lines
  }
}
\value{
  Constructs a contour plot of the launch angles
}
\author{
  Jim Albert
}


