\name{la_plot}
\alias{la_plot}

\title{
  Launch Angle Plot
}
\description{
  Constructs launch angle graph
}
\usage{
  la_plot(data, title)
}
\arguments{
  \item{data}{
    data frame containing Statcast data
  }
  \item{title}{
    title for the graph
  }
}
\value{
  Constructs a tile plot of the launch angles
}
\author{
  Jim Albert
}


