\name{ls_plot}
\alias{ls_plot}

\title{
  Launch Speed Plot
}
\description{
  Constructs launch speed graph
}
\usage{
  ls_plot(data, title)
}
\arguments{
  \item{data}{
    data frame containing Statcast data
  }
  \item{title}{
    title for the graph
  }
}
\value{
  Constructs a tile plot of the launch speeds
}
\author{
  Jim Albert
}


