\name{miss_swing_contour}
\alias{miss_swing_contour}

\title{
  Miss on Swings Contour Plot
}
\description{
  Constructs miss on swung balls contour graph
}
\usage{
  miss_swing_contour(data, L)
}
\arguments{
  \item{data}{
    data frame containing Statcast data
  }
  \item{L}{
    values of contour values
  }
}
\value{
  Constructs a contour plot of the miss on swung balls probability
}
\author{
  Jim Albert
}


