\name{miss_swing_plot}
\alias{miss_swing_plot}

\title{
  Miss on Swings Plot
}
\description{
  Constructs miss on swung balls graph
}
\usage{
  miss_swing_plot(df, title = "Probability of Miss", NCOL = 2)
}
\arguments{
  \item{df}{
    data frame or list containing Statcast data
  }
  \item{title}{
    title for the graph
  }
  \item{NCOL}{
    number of columns in multipanel display
  }
}
\value{
  Constructs a tile plot of the miss on swung balls probability
}
\author{
  Jim Albert
}


