\name{sa_plot}
\alias{sa_plot}

\title{
  Spray Angle Plot
}
\description{
  Constructs spray angle graph
}
\usage{
  sa_plot(df, title = "Spray Angle", NCOL = 2)
}
\arguments{
  \item{df}{
    data frame or list containing Statcast data
  }
  \item{title}{
    title for the graph
  }
  \item{NCOL}{
    number of columns in multipanel display
  }
}
\value{
  Constructs a tile plot of the spray angles
}
\author{
  Jim Albert
}


