\name{swing_contour}
\alias{swing_contour}

\title{
  Swing Contour Plot
}
\description{
  Constructs swing contour plot for a specific player
}
\usage{
  swing_contour(data, L)
}
\arguments{
  \item{data}{
    data frame containing Statcast data
  }
  \item{L}{
    values for contour lines
  }
}
\value{
  Constructs a contour plot of the swing probability
}
\author{
  Jim Albert
}


