\name{swing_plot}
\alias{swing_plot}

\title{
  Swing Plot
}
\description{
  Constructs swing plot for a specific player
}
\usage{
  swing_plot(data, title)
}
\arguments{
  \item{data}{
    data frame containing Statcast data
  }
  \item{title}{
    title for the graph
  }
}
\value{
  Constructs a tile plot of the swing probability
}
\author{
  Jim Albert
}


