\name{swing_plot}
\alias{swing_plot}

\title{
  Swing Plot
}
\description{
  Constructs swing plot
}
\usage{
  swing_plot(df, title = "Probability of Swing", NCOL = 2)
}
\arguments{
  \item{df}{
    data frame or list containing Statcast data
  }
  \item{title}{
    title for the graph
  }
  \item{NCOL}{
    number of columns in multipanel display
  }
}
\value{
  Constructs a tile plot of the swing probability
}
\author{
  Jim Albert
}


