\name{tile_plot}
\alias{tile_plot}

\title{
  Constructs heat map plot from GAM fit
}
\description{
  Constructs heat map plot from output from an generalized additive fit
}
\usage{
  tile_plot(fit, title)
}
\arguments{
  \item{fit}{
    saved gam fit
  }
  \item{title}{
    title for the graph
  }
}
\value{
  ggplot2 heat map graph with strike zone added
}
\author{
  Jim Albert
}


