\name{woba_contour}
\alias{woba_contour}

\title{
  wOBA Contour Plot
}
\description{
  Constructs wOBA angle contour graph
}
\usage{
  woba_contour(data, L, title)
}
\arguments{
  \item{data}{
    data frame containing Statcast data
  }
  \item{L}{
    values for contour lines
  }
  \item{title}{
    title of graph
  }
}
\value{
  Constructs a contour plot of the launch angles
}
\author{
  Jim Albert
}


