\name{woba_plot}
\alias{woba_plot}

\title{
  wOBA Plot
}
\description{
  Constructs wOBA graph
}
\usage{
  woba_plot(df, title = "Average wOBA", NCOL = 2)
}
\arguments{
  \item{df}{
    data frame or list containing Statcast data
  }
  \item{title}{
    title for the graph
  }
  \item{NCOL}{
    number of columns in multipanel display
  }
}
\value{
  Constructs a tile plot of the launch angles
}
\author{
  Jim Albert
}


